requirement.txt
torch
torchvision
einops
numpy
Pillow
tqdm
